Makefile: Set `PLATFORM_VERSION` value
`PLATFORM_VERSION` is a variable which contains the number of Android
version we're building Kernel for. For example, if we're building Kernel
for Android 10, then `PLATFORM_VERSION` is set to `10`.

`PLATFORM_VERSION` is used to decide correct drivers/configurations for
Android during build, e.g. to choose correct sdcardfs driver
(Ref: fs/sdcardfs/Kconfig).

Set `PLATFORM_VERSION` to `10`, as we're building Kernel for Android 10.

Signed-off-by: Hendra Manudinata <hendra@manudinata.me>