init: Fix RKP `vmm` compilation
To make `vmm.elf` file available to include during build,
it needs to be converted to IHEX, so it will be re-converted
back to its original format and copied to output directory.

Convert `vmm.elf` to IHEX and set up Makefile rule to
convert it back to ELF file to fix compilation issue.

Signed-off-by: Hendra Manudinata <hendra@manudinata.me>